docs: add LineChartWithXAxisPadding example in TypeScript (.tsx)
